+++ .gitignore
@@ -1,21 +1,6 @@
 target/
-pom.xml.tag
-pom.xml.releaseBackup
-pom.xml.versionsBackup
-pom.xml.next
-release.properties
 dependency-reduced-pom.xml
-buildNumber.properties
-.mvn/timing.properties
-.mvn/wrapper/maven-wrapper.jar
-
-# IntelliJ IDEA project files
 *.iml
 *.iws
 *.ipr
 .idea
-
-# Eclipse project files
-.settings
-.classpath
-.project
